Backend/requirements.txt
# Python 3.14–compatible versions (pandas 2.3.3+ has 3.14 wheels; numpy/scipy/sklearn support 3.14)
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Bcrypt>=1.0.1
Flask-JWT-Extended>=4.6.0
pymongo>=4.10.0
certifi>=2024.0.0
python-dotenv>=1.0.0
numpy>=2.0.0,<3
pandas>=2.3.3
scikit-learn==1.6.1
scipy>=1.14.0
hmmlearn>=0.3.2
joblib>=1.4.0
python-dateutil>=2.8.2
pytz>=2024.1
gunicorn>=23.0.0
sklearn-crfsuite>=0.5.0